Move color types/aliases to arcade.types.color
